requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
certifi==2021.5.30
charset-normalizer==2.0.6; python_version >= '3'
click==8.0.1
flask==2.0.1
idna==3.2; python_version >= '3'
importlib-metadata==4.8.1; python_version >= '3.6'
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.1; python_version >= '3.6'
keyring==23.2.1
markupsafe==2.0.1; python_version >= '3.6'
proxy-tools==0.1.0
pyobjc-core==7.3
pyobjc-framework-cocoa==7.3
pyobjc-framework-webkit==7.3; python_version >= '3.6'
pywebview==3.5
requests==2.26.0
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==2.0.1; python_version >= '3.6'
zipp==3.5.1; python_version >= '3.6'
